matchers.myers: Make node array a native C array of pointers
This is good for another order-of-magnitude speedup, but more
importantly is necessary to be able to release the GIL.

Breaking out node creation to the diagonal helper really makes
everything nicer, and as a bonus means we're only doing C-style
allocation and passing in the one place.

It's worth noting that we're using libc malloc here instead of the
Python-native alloc helpers, because they require the GIL and I'm
pretty keen on being able to drop that soon.